% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toydata.R
\docType{data}
\name{toydata}
\alias{toydata}
\title{Example data}
\format{
A data frame with 100 rows and 3 variables:
\describe{
  \item{y}{a binary outcome}
  \item{x1}{a continous predictor}
  \item{x2}{a binary predictor}
}
}
\source{
From CHL7001 Direct Reading Course: Statistical Programming and Computation for Health Data
}
\usage{
toydata
}
\description{
Data from a lecture example
}
\references{
course website
}
\keyword{datasets}
