% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funORCI.R
\name{OR_95CI}
\alias{OR_95CI}
\title{Create Odds Ratio and Confidence Interval}
\usage{
OR_95CI(coef, se, siglevel, roundto)
}
\arguments{
\item{coef}{coefficient of one variable}

\item{se}{standard deviation of one variable}

\item{siglevel}{significant level for confidence interval, range from 0 to 1}

\item{roundto}{a numeric value indicate the obtained decimal length}
}
\value{
ORresult
}
\description{
A function that returns the odds ratio and confidence interval
}
\examples{
logit_reg <- glm(y ~ x1 + x2, data = toydata, family = binomial(link = "logit"))
## FOR x1 variable
OR_95CI(summary(logit_reg)$coef[2,1],summary(logit_reg)$coef[2,2],0.95,2)
}
\author{
Yushu Zou
}
